\name{gcHFW}
\alias{gcHFW}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ garbage collection in global  environment }
\description{
  performs garbage collection in the global environment
}
\usage{
gcHFW()
}
%- maybe also 'usage' for other objects documented here.
\details{
  takes a sizable amount of time (ie is slow) but may be used for processes
  which take a lot of memory
}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of 'comp1'}
  \item{comp2 }{Description of 'comp2'}
  ...
}
\references{ ~put references to the literature/web site here ~ }
\author{ Felix Wittmann \email{hfwittmann@gmail.com}}

\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{

gcHFW()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip}
