\name{inst}
\alias{inst}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{provides path to working directory}
\description{
  if directory name ends with ~.Rcheck switch to sub-directory
  ~ contained in ~.Rcheck. Example directory name is  TAutils.Rcheck 
  then switch to sub-directory TAutils. 
}
\usage{
inst()
}
%- maybe also 'usage' for other objects documented here.
\details{
  useful for check process. In the check process these directories have already 
  been created.
}

\references{ ~put references to the literature/web site here ~ }
\author{ Felix Wittmann \email{hfwittmann@gmail.com}}

\examples{


}
\keyword{database}% at least one, from doc/KEYWORDS
